+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased] - XXXX-XX-XX
 
+
+## [0.20.3] - 2026-07-28
+
 ### Fixed
 
 - [#849](https://github.com/owncloud/user_ldap/pull/849) - Confine the LDAP home folder naming rule to the data directory
@@ -237,7 +240,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Add hint for max search term length - [#105](https://github.com/owncloud/user_ldap/issues/105)
 - Allow proxy to check next server - [#101](https://github.com/owncloud/user_ldap/issues/101)
 
-[Unreleased]: https://github.com/owncloud/user_ldap/compare/v0.20.2..master
+[Unreleased]: https://github.com/owncloud/user_ldap/compare/v0.20.3..master
+[0.20.3]: https://github.com/owncloud/user_ldap/compare/v0.20.2..v0.20.3
 [0.20.2]: https://github.com/owncloud/user_ldap/compare/v0.20.1..v0.20.2
 [0.20.1]: https://github.com/owncloud/user_ldap/compare/v0.20.0..v0.20.1
 [0.20.0]: https://github.com/owncloud/user_ldap/compare/v0.19.1..v0.20.0